dom/media/mediasource/test/test_BufferedSeek_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seeking in buffered range</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/mp4");

  sb.appendBuffer(new Uint8Array(await fetchWithXHR("bipbop/bipbop2s.mp4")));

  const target = 1.3;

  await once(v, "loadedmetadata");
  ok(true, "received loadedmetadata");
  v.currentTime = target;

  let wasSeeking = false;

  v.addEventListener("seeking", () => {
    wasSeeking = true;
    is(v.currentTime, target, "Video currentTime at target");
  });

  await once(v, "seeked");
  ok(wasSeeking, "Received expected seeking and seeked events");
  is(v.currentTime, target, "Video currentTime at target");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
